backend/requirements-raspi.txt
# Raspberry Pi 5 requirements for streaming + YOLO segmentation + Supabase uploads

# Core runtime
numpy>=1.24.0
opencv-python>=4.8.0
python-socketio[client]>=5.11.0
requests>=2.31.0

# Ultralytics (YOLOv8/YOLO11). Pulls torch/torchvision via piwheels on Pi.
ultralytics>=8.2.0

# Optional: if you plan to call Supabase directly from the Pi (not needed when using Edge Function)
# supabase>=2.4.0

# Optional: performance/monitoring utilities
# psutil>=5.9.0

# Optional: use ONNX instead of PyTorch (advanced setup)
# onnxruntime>=1.17.0


